feat: finish comment rest api

+++ src/controllers/comment.js
@@ -6,29 +6,38 @@ import passport from "passport";
 export const get = [
     validator.idParamFactory("postId"),
     async (req, res) => {
-        res.json(await db.readComment(null, Number(req.params.postId)));
+        const { postId } = req.params;
+        res.json(await db.readComment(undefined, Number(postId)));
     },
 ];
 
 // gets specific comment from a post
 export const getId = [
+    validator.idParamFactory("postId"),
     validator.idParamFactory("commentId"),
     validator.validateResults,
     async (req, res) => {
-        res.json(await db.readComment(Number(req.params.commentId)));
+        const { commentId, postId } = req.params;
+        const comment = await db.readComment(Number(commentId), Number(postId));
+        if (comment.length === 0) {
+            return res.status(404).end();
+        }
+        res.json(comment[0]);
     },
 ];
 
-// requires being at least a user
 export const post = [
     passport.authenticate("jwt", { session: false }),
     validator.idParamFactory("postId"),
     validator.textBodyFactory("text", "Comment", 1, 500),
     validator.validateResults,
     async (req, res) => {
-        // TODO: get authorid from req.user if authentication passed
-        let authorId = 1;
+        const { id: authorId } = req.user;
         const { postId } = req.params;
+        const post = await db.readPostById(Number(postId));
+        if (!post) {
+            return res.status(404).end();
+        }
         const { text } = req.body;
         res.json(
             await db.createComment(Number(postId), Number(authorId), text),
@@ -36,32 +45,44 @@ export const post = [
     },
 ];
 
-// TODO: validate user
-// requires being at least a user
-// user can only edit their own
 export const putId = [
+    passport.authenticate("jwt", { session: false }),
+    validator.idParamFactory("postId"),
     validator.idParamFactory("commentId"),
     validator.textBodyFactory("text", "Comment", 1, 500),
-    validator.date,
     validator.validateResults,
     async (req, res) => {
-        // TODO: validate that user is editing their own comment
-        const { commentId } = req.params;
-        const { text, date } = req.body;
-        res.json(await db.updateComment(Number(commentId), text, date));
+        const { id: authorId } = req.user;
+        const { postId, commentId } = req.params;
+        const comment = await db.readComment(Number(commentId), Number(postId));
+        if (comment.length === 0) {
+            return res.status(404).end();
+        }
+        if (comment[0].authorId !== authorId) {
+            return res.status(401).end();
+        }
+        const { text } = req.body;
+        if (text === comment[0].text) {
+            return res.json(comment[0]);
+        }
+        res.json(await db.updateComment(Number(commentId), text));
     },
 ];
 
-// TODO: validate user
-// user can only delete their own messages
-// an admin can delete any user's messages
 export const delId = [
+    passport.authenticate("jwt", { session: false }),
     validator.idParamFactory("commentId"),
     validator.validateResults,
     async (req, res) => {
-        // TODO: validate that user is deleting their own comment
-        // or that user is an admin
-        const { commentId } = req.params;
+        const { id: authorId, isAdmin } = req.user;
+        const { postId, commentId } = req.params;
+        const comment = await db.readComment(Number(commentId), Number(postId));
+        if (comment.length === 0) {
+            return res.status(404).end();
+        }
+        if (comment[0].authorId !== authorId && !isAdmin) {
+            return res.status(401).end();
+        }
         res.json(await db.deleteComment(Number(commentId)));
     },
 ];
